+++ src/helpers/isValidPost.ts
@@ -1,5 +1,5 @@
 import { measureChars } from '@/helpers/chars.js';
-import { isValidPoll } from '@/helpers/createPoll.js';
+import { isValidPoll } from '@/helpers/polls.js';
 import { getCurrentPostLimits } from '@/helpers/getCurrentPostLimits.js';
 import type { CompositePost } from '@/store/useComposeStore.js';
