+++ src/helpers/resolveSimpleHashChain.ts
@@ -1,16 +1,17 @@
 /* cspell:disable */
 
 import { ChainId } from '@masknet/web3-shared-evm';
+import { first, memoize } from 'lodash-es';
 
 const EVM_CHAIN: { [key in ChainId]?: string } = {
     [ChainId.Mainnet]: 'ethereum',
+    [ChainId.Base]: 'base',
     [ChainId.BSC]: 'bsc',
     [ChainId.Matic]: 'polygon',
     [ChainId.Arbitrum]: 'arbitrum',
     [ChainId.Optimism]: 'optimism',
     [ChainId.Avalanche]: 'avalanche',
     [ChainId.xDai]: 'gnosis',
-    [ChainId.Base]: 'base',
     [ChainId.Scroll]: 'scroll',
     [ChainId.Celo]: 'celo',
     [ChainId.Zora]: 'zora',
@@ -21,3 +22,9 @@ const EVM_CHAIN: { [key in ChainId]?: string } = {
 export function resolveSimpleHashChain(chain: ChainId) {
     return EVM_CHAIN[chain] ?? undefined;
 }
+
+export const resolveSimpleHashChainId: (chainId: string) => ChainId | undefined = memoize(function resolveChainId(
+    chain: string,
+): ChainId | undefined {
+    return first(Object.entries(EVM_CHAIN).find(([, value]) => value === chain)) as ChainId | undefined;
+});
